Add makeMovie.m
The makeMovie() function allows for a MATLAB movie to be created out
of the images in a given directory. The function utilizes the Parallel
Computing Toolbox to speed up processing time. Currently, the function
assumes that RemoveXRays() must be called to process the images, but
this can be changed in the future.