Add app_data variable and make other major changes
Add two new classes, AppData and DatasetInfo, that will keep track of
information that used to be loaded/stored using setappdata() and
getappdata(). The purpose of this is two-fold: First, it allows other
functions to work on things in advance. What this means is that the
functions do not have to use the current state of the program to do
their processing on (such as only being able to load an image that
match the current images/stack values). Now, functions can be passed
a modified version of app_data that gives them the appropriate
information to work on a different "state" of the program (e.g. work
on loading an image with different image/stack values than the current
ones). Second, changes to processed_data (which is now stored in app_data)
should be accessible in real-time, as app_data is a handle object. This
is necessary in order to see changes to processed_data that come about
due to parallel processing.

Other changes include converting scripts into functions, so that they
can work on different "states" of the program (as described in the
previous paragraph).