Add cropping ability
Add the ability to crop and uncrop an image, effectively setting a
region of interest (ROI) through which other images can be viewed.
The crop() and uncrop() functions can be called from the MATLAB
command window to achieve this functionality.

A known issue with this version is incorrectly adjusting energy tick
marks on the y-axis when cropping images from energy cameras.